requirements.txt
streamlit
requests
python-whois
base64
socket
re
